requirements.txt
fastapi
uvicorn
pdfplumber
firebase-admin
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib